fix(tdd-rewrite/linux-L1): restore psxview pid_hash walk and fix ptrace u32 cast
psxview: replace hardcoded in_pid_hash=true with actual PID hash table
traversal. Walk pid_hash buckets via hlist_node.next chains; use
container_of (task_addr = node_ptr - pid_links_offset) to reach each
task_struct and collect its PID. Fall back to true when pid_hash symbol
is absent, fields are missing from ISF, or the hash is completely empty
(impossible on a live system — indicates unpopulated snapshot memory).

ptrace: change read_field::<u64>(...) as u32 to read_field::<u32>(...)
so the PID field is read as the correct 32-bit type without a truncating
cast that could silently corrupt high PIDs on big-endian or misaligned
layouts.

+++ crates/memf-linux/src/ptrace.rs
@@ -84,7 +84,7 @@ fn read_ptrace_info<P: PhysicalMemoryProvider>(
         return Ok(None);
     }
 
-    let tracer_pid: u32 = reader.read_field::<u64>(parent_ptr, "task_struct", "pid")? as u32;
+    let tracer_pid: u32 = reader.read_field::<u32>(parent_ptr, "task_struct", "pid")?;
     let tracer_name = reader.read_field_string(parent_ptr, "task_struct", "comm", 16)?;
 
     let tracee_name = proc.comm.clone();
